+++ src/main/java/vn/edu/hcmuaf/fit/websubject/repository/CategoryRepository.java
@@ -2,13 +2,13 @@ package vn.edu.hcmuaf.fit.websubject.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import vn.edu.hcmuaf.fit.websubject.model.Category;
+import vn.edu.hcmuaf.fit.websubject.entity.Category;
 
 import java.util.List;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-    List<Category> findByParentCategoryId(Long parentId);
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+    List<Category> findByParentCategoryId(Integer parentId);
 
     List<Category> findByParentCategoryIsNull();
 }
