+++ src/main/java/vn/edu/hcmuaf/fit/websubject/repository/ProductRepository.java
@@ -3,12 +3,12 @@ package vn.edu.hcmuaf.fit.websubject.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import vn.edu.hcmuaf.fit.websubject.model.Product;
+import vn.edu.hcmuaf.fit.websubject.entity.Product;
 
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
     @Query("SELECT p FROM Product p WHERE p.category.id = :categoryId OR p.category.parentCategory.id = :categoryId")
-    List<Product> findByCategoryParentOrCategory(Long categoryId);
+    List<Product> findByCategoryParentOrCategory(Integer categoryId);
 }
